Convert all utils to TS
